# soldout/api/openapi.yaml
openapi: 3.1.0
info:
  title: soldout API
  version: "1.2.0-hw1-v2"
  description: |
    HTTP API системы продажи билетов `soldout` (модульный монолит, занятие 1).
    Аутентификации нет (появится на занятии 6 через Gravitee + Keycloak). Все мутирующие операции
    с внешним эффектом идемпотентны: `POST /v1/orders` — по заголовку `Idempotency-Key`,
    `POST /v1/orders/{id}/pay` — по заказу (повтор возвращает тот же результат).
servers:
  - url: http://localhost:8080
tags:
  - name: catalog
  - name: queue
  - name: booking
  - name: ticketing
  - name: admin
  - name: platform

paths:
  /v1/events:
    get:
      tags: [catalog]
      summary: Список мероприятий
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: { type: array, items: { $ref: "#/components/schemas/Event" } }

  /v1/events/{id}:
    get:
      tags: [catalog]
      summary: Мероприятие
      parameters: [{ $ref: "#/components/parameters/UUIDPath" }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Event" } } } }
        "404": { $ref: "#/components/responses/Error" }

  /v1/events/{id}/seatmap:
    get:
      tags: [catalog]
      summary: Сводка карты зала по секторам
      description: |
        С занятия 2 — счётчики free/held/sold и версия карты каждого сектора из проекции `catalog_seat_state`
        (≈ 3 КБ вместо 1.8 МБ). Полная карта — по секторам, см. `/sectors/{sector_id}/seatmap`.
      parameters: [{ $ref: "#/components/parameters/UUIDPath" }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SeatMapSummary" } } } }
        "404": { $ref: "#/components/responses/Error" }

  /v1/events/{id}/sectors/{sector_id}/seatmap:
    get:
      tags: [catalog]
      summary: Карта одного сектора (≈ 1 000 мест, ≈ 32 КБ)
      description: |
        Предсобранный JSON из кэша (Valkey) с версией из БД. `ETag: "<version>"`; при `If-None-Match` с той же версией — 304.
        Заголовок `X-Cache: hit|miss` — для демонстрации.
      parameters:
        - { $ref: "#/components/parameters/UUIDPath" }
        - name: sector_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: If-None-Match
          in: header
          required: false
          schema: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SectorSeatMap" } } } }
        "304": { description: Не изменилось }
        "404": { $ref: "#/components/responses/Error" }

  /v1/events/{id}/sectors/{sector_id}/price:
    get:
      tags: [booking]
      summary: Цена следующего места сектора с динамической ценой (HW1, V2)
      description: |
        Цена k-го удержанного или проданного места сектора: `base * (1 + step_pct/100)^floor((k-1)/step_seats)`,
        округление вниз до копейки. `next_ordinal` = число active и confirmed hold'ов сектора + 1.
        Цена справочная: hold фиксирует цену своего порядкового номера в момент создания.
      parameters:
        - { $ref: "#/components/parameters/UUIDPath" }
        - name: sector_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SectorPrice" } } } }
        "404": { description: У сектора нет динамической цены (`no_pricing`) }

  /v1/events/{id}/seatmap/stream:
    get:
      tags: [catalog]
      summary: SSE-поток изменений статусов мест (появляется на занятии 2)
      parameters: [{ $ref: "#/components/parameters/UUIDPath" }]
      responses:
        "501": { description: Не реализовано на занятии 1 }

  /v1/events/{id}/queue/join:
    post:
      tags: [queue]
      summary: Войти в waiting room
      description: На занятии 1 всегда выдаёт токен допуска (position = 0). Токен передаётся в `X-Admission-Token` при создании hold.
      parameters: [{ $ref: "#/components/parameters/UUIDPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, format: uuid }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Admission" } } } }
        "404": { $ref: "#/components/responses/Error" }

  /v1/holds:
    post:
      tags: [booking]
      summary: Удержать место (hold на 10 минут)
      description: |
        В секторе с динамической ценой (HW1, V2) hold получает порядковый номер в секторе (`ordinal`) и цену этого номера
        (`price_minor`); цена не меняется до оплаты. Сумма заказа равна сумме цен его hold'ов.
      parameters:
        - name: X-Admission-Token
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_id, seat_id, user_id]
              properties:
                event_id: { type: string, format: uuid }
                seat_id: { type: integer, format: int64 }
                user_id: { type: string, format: uuid }
      responses:
        "201": { description: Создано, content: { application/json: { schema: { $ref: "#/components/schemas/Hold" } } } }
        "401": { description: Нет/недействителен токен допуска (`admission_required`) }
        "404": { $ref: "#/components/responses/Error" }
        "409": { description: Место занято (`seat_held`), продано (`seat_sold`) или продажи закрыты (`sales_closed`) }
        "422": { description: Лимит 4 билета (`hold_limit`), место не с этой площадки (`seat_not_in_event`), валидация }

  /v1/holds/{id}:
    get:
      tags: [booking]
      summary: Удержание
      parameters: [{ $ref: "#/components/parameters/UUIDPath" }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Hold" } } } }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [booking]
      summary: Снять удержание
      parameters: [{ $ref: "#/components/parameters/UUIDPath" }]
      responses:
        "204": { description: Снято }
        "404": { $ref: "#/components/responses/Error" }
        "409": { description: Hold не активен (`hold_state`) }

  /v1/orders:
    post:
      tags: [booking]
      summary: Создать заказ из hold'ов
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hold_ids, user_id]
              properties:
                hold_ids: { type: array, items: { type: string, format: uuid }, minItems: 1, maxItems: 4 }
                user_id: { type: string, format: uuid }
      responses:
        "201": { description: Создан, content: { application/json: { schema: { $ref: "#/components/schemas/Order" } } } }
        "200": { description: Повтор с тем же Idempotency-Key — существующий заказ }
        "404": { $ref: "#/components/responses/Error" }
        "409": { description: Hold истёк/не активен/уже в заказе (`hold_state`), конфликт ключа (`idempotency_conflict`) }
        "422": { description: Валидация, лимит 4, hold'ы разных пользователей/мероприятий }

  /v1/orders/{id}:
    get:
      tags: [booking]
      summary: Заказ
      parameters: [{ $ref: "#/components/parameters/UUIDPath" }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Order" } } } }
        "404": { $ref: "#/components/responses/Error" }

  /v1/orders/{id}/pay:
    post:
      tags: [booking]
      summary: Оплатить заказ
      description: |
        Вызывает платёжный эмулятор (`PSP_FAIL_RATE`, `PSP_DELAY_MS`). При успехе hold'ы подтверждаются,
        синхронно выпускаются билеты и уведомление (занятие 1); заказ → `paid`. При отказе hold'ы освобождаются, заказ → `failed`.
        Повторный вызов для оплаченного заказа идемпотентен.
      parameters:
        - { $ref: "#/components/parameters/UUIDPath" }
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      responses:
        "200": { description: Оплачен, content: { application/json: { schema: { $ref: "#/components/schemas/Order" } } } }
        "402": { description: Оплата отклонена (`payment_failed`) }
        "404": { $ref: "#/components/responses/Error" }
        "409": { description: Заказ истёк (`order_expired`) или не pending (`order_state`) }

  /v1/users/{id}/tickets:
    get:
      tags: [ticketing]
      summary: Билеты пользователя
      parameters: [{ $ref: "#/components/parameters/UUIDPath" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets: { type: array, items: { $ref: "#/components/schemas/Ticket" } }

  /v1/admin/events:
    post:
      tags: [admin]
      summary: Создать мероприятие (без auth на занятии 1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, venue_id, starts_at, sales_open_at]
              properties:
                name: { type: string }
                venue_id: { type: string, format: uuid }
                starts_at: { type: string, format: date-time }
                sales_open_at: { type: string, format: date-time }
      responses:
        "201": { description: Создано, content: { application/json: { schema: { $ref: "#/components/schemas/Event" } } } }
        "422": { $ref: "#/components/responses/Error" }

  /v1/admin/events/{id}/open:
    post:
      tags: [admin]
      summary: Открыть продажи
      parameters: [{ $ref: "#/components/parameters/UUIDPath" }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Event" } } } }
        "404": { $ref: "#/components/responses/Error" }
        "409": { description: Недопустимый переход (`invalid_transition`) }

  /healthz:
    get: { tags: [platform], summary: Процесс жив, responses: { "200": { description: OK } } }
  /readyz:
    get: { tags: [platform], summary: Зависимости отвечают, responses: { "200": { description: OK }, "503": { description: Не готов } } }
  /metrics:
    get: { tags: [platform], summary: Prometheus-метрики, responses: { "200": { description: OK } } }
  /debug/pprof/{profile}:
    get:
      tags: [platform]
      summary: Профили Go (pprof)
      parameters: [{ name: profile, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }

components:
  parameters:
    UUIDPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: seat_held }
            message: { type: string }
    Event:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        venue_id: { type: string, format: uuid }
        starts_at: { type: string, format: date-time }
        sales_open_at: { type: string, format: date-time }
        sales_state: { type: string, enum: [scheduled, open, closed] }
    SeatMapSummary:
      type: object
      properties:
        event_id: { type: string, format: uuid }
        sectors:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              name: { type: string }
              kind: { type: string, enum: [seated, standing] }
              total: { type: integer }
              free: { type: integer }
              held: { type: integer }
              sold: { type: integer }
              version: { type: integer, format: int64 }
    SectorSeatMap:
      type: object
      properties:
        event_id: { type: string, format: uuid }
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string, enum: [seated, standing] }
        version: { type: integer, format: int64 }
        seats:
          type: array
          items:
            type: object
            properties:
              id: { type: integer, format: int64 }
              row: { type: integer }
              num: { type: integer }
              status: { type: string, enum: [free, held, sold] }
    Admission:
      type: object
      properties:
        token: { type: string }
        event_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        position: { type: integer }
        issued_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
    Hold:
      type: object
      properties:
        id: { type: string, format: uuid }
        event_id: { type: string, format: uuid }
        seat_id: { type: integer, format: int64 }
        user_id: { type: string, format: uuid }
        status: { type: string, enum: [active, released, confirmed] }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        price_minor: { type: integer, format: int64, description: "цена в копейках, зафиксированная в момент hold (только сектор с динамической ценой)" }
        ordinal: { type: integer, description: "порядковый номер hold в секторе (только сектор с динамической ценой)" }
    SectorPrice:
      type: object
      properties:
        event_id: { type: string, format: uuid }
        sector_id: { type: string, format: uuid }
        next_ordinal: { type: integer }
        price_minor: { type: integer, format: int64 }
        price: { type: string, example: "5250.00" }
        base_price_minor: { type: integer, format: int64 }
        step_seats: { type: integer, example: 200 }
        step_pct: { type: integer, example: 5 }
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        event_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        status: { type: string, enum: [pending, paid, failed, expired] }
        amount: { type: string, example: "5000.00" }
        amount_minor: { type: integer, format: int64, description: "сумма цен hold'ов заказа в копейках" }
        hold_ids: { type: array, items: { type: string, format: uuid } }
        seat_ids: { type: array, items: { type: integer, format: int64 } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Ticket:
      type: object
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        seat_id: { type: integer, format: int64 }
        code: { type: string }
        issued_at: { type: string, format: date-time }